refactor: replace old switch with design system switch
