--- src/components/Switch/Switch.md
@@ -1,5 +1,0 @@
-```jsx
-const [isChecked, setIsChecked] = React.useState(false);
-
-<Switch checked={isChecked} onChange={() => setIsChecked(!isChecked)} />;
-```
