--- src/components/Switch/index.ts
@@ -1,3 +1,0 @@
-import Switch from './Switch';
-
-export default Switch;
